moved look around into interact.js
