+++ server/crud/interact/interact.js
@@ -2,11 +2,13 @@ const crud_user_basic = require('../user/basic');
 const crud_connection = require('../user/connection');
 const proximity = require('./proximity');
 const crud_attack = require('./attack');
+const crud_terrain = require('../terrain')
 const config = require("../../config");
 
 module.exports = {
     announce,
-    attack_nearby
+    attack_nearby,
+    look_around
 }
 
 function announce(socket_id, io, message, distance, check_behind) {
@@ -67,5 +69,33 @@ function attack_nearby(socket, io, distance, energy, damage, check_behind) {
                 }
             });
         });
+    });
+}
+
+function look_around(socket_id, io) {
+    
+    crud_user_basic.get_user(socket_id).catch(console.dir).then( (user) => {
+        // display biomes nearby
+        crud_terrain.check_biomes(socket_id, io, user["angle"], user["lat"], user["long"]);
+
+        // display other players nearby
+        crud_connection.get_other_connections(
+            socket_id, user["lat"], user["long"], config.ONE_METER*config.SEEING_DISTANCE
+        ).catch(console.dir).then( (other_users) => {
+            other_users.forEach( (other_user) => {
+                // check if "user" can see "other_user"
+                var perspective = proximity.get_perspective(
+                    other_user, user, config.ONE_METER*config.SEEING_DISTANCE, true
+                );
+                if(perspective) {
+                    io.to(socket_id).emit('message', {
+                        data: 'You see a ' + other_user['tall'] + ' ' + other_user['weight'] + 
+                        ' ' + other_user['age'] + ' human ' +  perspective
+                    });
+                }
+            })
+        });
+
+        // display corpses nearby
     });
 }
