+++ server/router/login.js
@@ -1,7 +1,7 @@
 const crud_user = require('../crud/user/user');
 const crud_user_basic = require('../crud/user/basic');
 const crud_connection = require('../crud/user/connection');
-const crud_look_around = require('../crud/interact/look_around');
+const crud_interact = require('../crud/interact/interact');
 const crud_patch_notes = require('../crud/patch_notes');
 const { write_patch_notes } = require('./patch_notes');
 
@@ -24,7 +24,8 @@ function login(data, io, socket) {
 
     crud_user.create_user(data["email"], socket).catch(console.dir).then( () => {
         crud_connection.add_connection(data["email"], socket.id).then( () => {
-            crud_look_around.look_around(socket.id, io);
+            
+crud_interact.look_around(socket.id, io);
             crud_user_basic.get_user(socket.id).catch(console.dir).then( (user) => {
                 crud_patch_notes.get_patch_notes_since_ts(
                     user['last_read_patch_notes']
