+++ server/router/look.js
@@ -1,6 +1,6 @@
 const crud_move = require('../crud/move');
 const crud_user_basic = require('../crud/user/basic');
-const crud_look_around = require('../crud/interact/look_around');
+const crud_interact = require('../crud/interact/interact');
 
 module.exports = {
     add_routes
@@ -45,6 +45,7 @@ function add_routes(socket, io) {
     })
 
     socket.on('look', function(data) {
-        crud_look_around.look_around(socket.id, io);
+        
+crud_interact.look_around(socket.id, io);
     })
 }
